refactor test vectors
