+++ packages/_test-vectors/index.ts
@@ -1,0 +1,2 @@
+export * from "./mockedBoxes";
+export * from "./mockedTransactions";
