+++ packages/_test-vectors/mockedBoxes.ts
@@ -1,6 +1,6 @@
 import { Amount, Box } from "packages/common/src/";
 
-export const regularBoxesMock: Box<bigint>[] = [
+export const regularBoxes: Box<bigint>[] = [
   {
     boxId: "e56847ed19b3dc6b72828fcfb992fdf7310828cf291221269b7ffc72fd66706e",
     value: 67500000000n,
@@ -168,7 +168,7 @@ export const regularBoxesMock: Box<bigint>[] = [
   }
 ];
 
-export const manyTokensBoxesMock: Box<Amount>[] = [
+export const manyTokensBoxes: Box<Amount>[] = [
   {
     boxId: "490148afdc36f5459bbfd84922a446abea9a1077e031822f377b0ff3a6e467e3",
     transactionId: "c82615aa845d8159b7a9e33401c0d4c56535a8f40c3b40b4d86fcbc15084bf0f",
@@ -1300,7 +1300,7 @@ export const manyTokensBoxesMock: Box<Amount>[] = [
   }
 ];
 
-export const validBoxesMock = [
+export const validBoxes = [
   {
     boxId: "135baecae94f7ec20caf981800166d450bd1dde4b959e5fdd0e2751b679d94dd",
     transactionId: "ae11d207f0989945f63909d2f703b2640acf4f654a8fdadd23570a640f9d12ee",
@@ -1498,7 +1498,7 @@ export const validBoxesMock = [
   }
 ];
 
-export const invalidBoxesMock = [
+export const invalidBoxes = [
   {
     boxId: "135baecae94f7ec21caf981800166d450bd1dde4b959e5fdd0e2751b679d94dd",
     transactionId: "ae11d207f0989945f63909d2f703b2640acf4f654a8fdadd23570a640f9d12ee",
@@ -1718,3 +1718,271 @@ export const outOfOrderRegistersBox = {
   transactionId: "63661fccf4621fc9c9864f2cf5bb6a14ff7784156504101ac4199dd400196d82",
   index: 0
 };
+
+export const babelBoxesMock: Box<string>[] = [
+  {
+    boxId: "8fccab355e7b02f758f4aeab625a8b0b42c75a21c9855d3a40d9bb3e06ba70d7",
+    transactionId: "b34bf160ed1d14185884836ccbbc4932b4cccf13855321883a1a695837d5fce8",
+    value: "347000000",
+    creationHeight: 894206,
+    index: 1,
+    ergoTree:
+      "100604000e2003faf2cb329f2e90d6d23b58d91bbb6c046aa143261cc21f52fbe2824bfcbf040400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd0236b7247e35a4431813623360165dbd5209c2fdec0618fe114c74966e2cded785",
+      R5: "0580a4e803",
+      R6: "0e20a47d86185d544ec6ff0fee7efd7a84ac430704bd6631b92c69543079445a5053"
+    },
+    assets: [
+      {
+        tokenId: "03faf2cb329f2e90d6d23b58d91bbb6c046aa143261cc21f52fbe2824bfcbf04",
+        amount: "16"
+      }
+    ]
+  },
+  {
+    boxId: "0f26f606aa0480739f7eebdda0ed61755d85791abe0d2ecc1ce06d12acb65dcd",
+    transactionId: "cd5c73e7b68b25c9f0e3a2449209abc78a2d16abc0d2d044e7e5e3e2d4034a08",
+    value: "990000000",
+    creationHeight: 892499,
+    index: 1,
+    ergoTree:
+      "100604000e20472c3d4ecaa08fb7392ff041ee2e6af75f4a558810a74b28600549d5392810e80400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd0377f4a3eb90abc59d3d31a0860bd9215444f467e743d9ef1de3262a1d35fa4ae5",
+      R5: "050a",
+      R6: "0e20b4ab781069c0f0a24d9a3097b5d5955a76c1dc72f217a21a25698f1c51ff2e55"
+    },
+    assets: [
+      {
+        tokenId: "472c3d4ecaa08fb7392ff041ee2e6af75f4a558810a74b28600549d5392810e8",
+        amount: "2200000"
+      }
+    ]
+  },
+  {
+    boxId: "af78e056fdc0fde1c32dd6ff4615bf346a7e73b164aeac519dd2a2fd583bdbf6",
+    transactionId: "794f023234fb9160c104dae57598ea428327806ea2cdd6b53538a9fb777b7144",
+    value: "900956118",
+    creationHeight: 892056,
+    index: 1,
+    ergoTree:
+      "100604000e203405d8f709a19479839597f9a22a7553bdfc1a590a427572787d7c44a88b63860400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd0205487349cb347ab113ae9edeb300974cb418f64197613021b1c0ba8648bc5125",
+      R5: "05ccf3e00b",
+      R6: "0e20d3c0bb4272f54b9642ebde66cab5bd553154a9504eb772d7341e7acfb6e3dd8b"
+    },
+    assets: [
+      {
+        tokenId: "3405d8f709a19479839597f9a22a7553bdfc1a590a427572787d7c44a88b6386",
+        amount: "4"
+      }
+    ]
+  },
+  {
+    boxId: "94a46d178733b0e94daa12ec740189ecb3be90c7b10429e24f93e7455d33f514",
+    transactionId: "f25927c75bb1957c91f4bb304e8b883b528120ab30d2cfcbc9a8bcb88ea24ee9",
+    value: "576712000",
+    creationHeight: 892037,
+    index: 1,
+    ergoTree:
+      "100604000e200cd8c9f416e5b1ca9f986a7f10a84191dfb85941619e49e53c0dc30ebf83324b0400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd03e4f04a919ad81b405d1634208416c76c9b188c210679433236792a3ba9468cf1",
+      R5: "05e05d",
+      R6: "0e20c1f34276b837401528f5be161364500860603d49f34af5942366bb8dcedace89"
+    },
+    assets: [
+      {
+        tokenId: "0cd8c9f416e5b1ca9f986a7f10a84191dfb85941619e49e53c0dc30ebf83324b",
+        amount: "4048"
+      }
+    ]
+  },
+  {
+    boxId: "444c0d19a63394dfcfd7fc74296e07415ce46fbd7be2f60bf76cc981131582d5",
+    transactionId: "4a86fbedafa665cb28e9b97ece5fb5c6c1c89f40c4769437e624f9d3739760b9",
+    value: "979000000",
+    creationHeight: 889772,
+    index: 1,
+    ergoTree:
+      "100604000e2000bd762484086cf560d3127eb53f0769d76244d9737636b2699d55c56cd470bf0400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd0377f4a3eb90abc59d3d31a0860bd9215444f467e743d9ef1de3262a1d35fa4ae5",
+      R5: "0514",
+      R6: "0e20068b7dac20044d27a6dff88cf1510f9ba64c0a83b5e3f5933253d1094e6af911"
+    },
+    assets: [
+      {
+        tokenId: "00bd762484086cf560d3127eb53f0769d76244d9737636b2699d55c56cd470bf",
+        amount: "2200000"
+      }
+    ]
+  },
+  {
+    boxId: "ef719d69ecf3d78b6d35f4de9efeb89c7296fceb2a7122b401e2051e036bbc2b",
+    transactionId: "32d8b9a6d045e464f82cfda5d119cb1152c91e38e439e8e546cf4f6c9834e297",
+    value: "7985320000",
+    creationHeight: 888882,
+    index: 1,
+    ergoTree:
+      "100604000e20d71693c49a84fbbecd4908c94813b46514b18b67a99952dc1e6e4791556de4130400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd0377f4a3eb90abc59d3d31a0860bd9215444f467e743d9ef1de3262a1d35fa4ae5",
+      R5: "0580e209",
+      R6: "0e20ef6772c4277811b42bcffdf2e3d6ec953b66983670419834003e96ba1a624caa"
+    },
+    assets: [
+      {
+        tokenId: "d71693c49a84fbbecd4908c94813b46514b18b67a99952dc1e6e4791556de413",
+        amount: "196"
+      }
+    ]
+  },
+  {
+    boxId: "bc3d6ba462cc04db5d7940078d3b07782ecb43f6b277e59d68d93ec412457302",
+    transactionId: "cf1bfd596542c53510ea4b9e93780a482d75505a3cec5c5c055273c935508657",
+    value: "996600000",
+    creationHeight: 883249,
+    index: 1,
+    ergoTree:
+      "100604000e2002f31739e2e4937bb9afb552943753d1e3e9cdd1a5e5661949cb0cef93f907ea0400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd0377f4a3eb90abc59d3d31a0860bd9215444f467e743d9ef1de3262a1d35fa4ae5",
+      R5: "0514",
+      R6: "0e20ddc68ae0b9bcd2c806185eeabf745bf61b2d7a3c4986a132ef78a8f0f742db1b"
+    },
+    assets: [
+      {
+        tokenId: "02f31739e2e4937bb9afb552943753d1e3e9cdd1a5e5661949cb0cef93f907ea",
+        amount: "440000"
+      }
+    ]
+  },
+  {
+    boxId: "556dfb6bc2d1d285a8996a588658eb3504a5c32de160edd16011f756599b347a",
+    transactionId: "eabb7db533448cb2a7d67cb937e6692670df33dedb0802449845eb304c98e48f",
+    value: "29997400000",
+    creationHeight: 880390,
+    index: 1,
+    ergoTree:
+      "100604000e20003bd19d0187117f130b62e1bcab0939929ff5c7709f843c5c4dd158949285d00400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd0377f4a3eb90abc59d3d31a0860bd9215444f467e743d9ef1de3262a1d35fa4ae5",
+      R5: "05c0cf24",
+      R6: "0e2042b83bc82d1a482b8c5975bdb4403a1c6cf2d54624e403188e2af9ac558414ed"
+    },
+    assets: [
+      {
+        tokenId: "003bd19d0187117f130b62e1bcab0939929ff5c7709f843c5c4dd158949285d0",
+        amount: "12"
+      }
+    ]
+  },
+  {
+    boxId: "008c7e08379833b3888be6f9d16c82bc58b91b936b2f0309b44226562feff73c",
+    transactionId: "e11c29597af4180fa540d80fe9ee4e370eecd547410692cebb5a488addc4ccb0",
+    value: "24996000000",
+    creationHeight: 874935,
+    index: 1,
+    ergoTree:
+      "100604000e2003faf2cb329f2e90d6d23b58d91bbb6c046aa143261cc21f52fbe2824bfcbf040400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd0377f4a3eb90abc59d3d31a0860bd9215444f467e743d9ef1de3262a1d35fa4ae5",
+      R5: "05c096b102",
+      R6: "0e2012cd9937256c16cc8796cb8dd19ef0ad61fa8ced3fe3eed2f28105b7db38724a"
+    },
+    assets: [
+      {
+        tokenId: "03faf2cb329f2e90d6d23b58d91bbb6c046aa143261cc21f52fbe2824bfcbf04",
+        amount: "2"
+      }
+    ]
+  },
+  {
+    boxId: "98df6f57cba1c1d1469388e3159132535891dcec64df5831284e48829be61fbf",
+    transactionId: "999cb7866e0bf99fb16eb6515f491cf90f59674c4b6dedd3526612d1e1ba3376",
+    value: "2496600000",
+    creationHeight: 874929,
+    index: 1,
+    ergoTree:
+      "100604000e20e8b20745ee9d18817305f32eb21015831a48f02d40980de6e849f886dca7f8070400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd0377f4a3eb90abc59d3d31a0860bd9215444f467e743d9ef1de3262a1d35fa4ae5",
+      R5: "0502",
+      R6: "0e201c7e5f7245d7ac19b078200985b989abb1b9dc42ac0ff486dc65677dd2980e2a"
+    },
+    assets: [
+      {
+        tokenId: "e8b20745ee9d18817305f32eb21015831a48f02d40980de6e849f886dca7f807",
+        amount: "4400000"
+      }
+    ]
+  },
+  {
+    boxId: "bf1085abdd2f2ba7c85002ff1ab1f80d4032ad03ff4aa042f5789b58a58c626d",
+    transactionId: "f6c88acb142719b988f058ea52a6cb2fe63195ba95aa1ac8508e6e7339323f2a",
+    value: "2998799980",
+    creationHeight: 874925,
+    index: 1,
+    ergoTree:
+      "100604000e2018c938e1924fc3eadc266e75ec02d81fe73b56e4e9f4e268dffffcb30387c42d0400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd0377f4a3eb90abc59d3d31a0860bd9215444f467e743d9ef1de3262a1d35fa4ae5",
+      R5: "053c",
+      R6: "0e20a3589ad74f0604775fcda94ef2e9c50b5825c04c0f6b55793c5a0af866228d63"
+    },
+    assets: [
+      {
+        tokenId: "18c938e1924fc3eadc266e75ec02d81fe73b56e4e9f4e268dffffcb30387c42d",
+        amount: "73334"
+      }
+    ]
+  },
+  {
+    boxId: "45b1cf96e1e25cac847334dbd972b1b7c9fa74dd276856e9ac40cd4f4c7485b1",
+    transactionId: "4123e5d83aadec1bc225ae52beeef2600354ecff3523ca0b9038ade88665190d",
+    value: "1001000000",
+    creationHeight: 873619,
+    index: 0,
+    ergoTree:
+      "100604000e203405d8f709a19479839597f9a22a7553bdfc1a590a427572787d7c44a88b63860400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd0377f4a3eb90abc59d3d31a0860bd9215444f467e743d9ef1de3262a1d35fa4ae5",
+      R5: "05c09a0c",
+      R6: "0e00"
+    },
+    assets: []
+  },
+  {
+    boxId: "fc3aa86922611be22ce5a1ab844add3c0b4f72aaee5657be3158c462e4e209fe",
+    transactionId: "9ff5c883edd67b0ce6c58817187b5da95285ab0a7c60fa5ab0d15b7a60dcbf4f",
+    value: "10001000000",
+    creationHeight: 873592,
+    index: 0,
+    ergoTree:
+      "100604000e2003faf2cb329f2e90d6d23b58d91bbb6c046aa143261cc21f52fbe2824bfcbf040400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd02a607f318d8b1e4cf636c9a1561c2f0e84be8cc15d008de54a757ac787b49b8ef",
+      R5: "0580897a",
+      R6: "0e00"
+    },
+    assets: []
+  },
+  {
+    boxId: "d52d58318a5e6b89a0976e4d53fae432321d7d2abb9399cdc8b38b2944db3dc5",
+    transactionId: "2ee82e6d1b018b8a08dc0a77de52fe01933741b61fcd9151d9b7c569449fff80",
+    value: "10001000000",
+    creationHeight: 873442,
+    index: 0,
+    ergoTree:
+      "100604000e20d71693c49a84fbbecd4908c94813b46514b18b67a99952dc1e6e4791556de4130400040005000500d803d601e30004d602e4c6a70408d603e4c6a7050595e67201d804d604b2a5e4720100d605b2db63087204730000d606db6308a7d60799c1a7c17204d1968302019683050193c27204c2a7938c720501730193e4c672040408720293e4c672040505720393e4c67204060ec5a796830201929c998c7205029591b1720673028cb272067303000273047203720792720773057202",
+    additionalRegisters: {
+      R4: "08cd0256212b499e64bae4696f21e995515a2f525e5e843eea62f0b8e7dd64f96988a7",
+      R5: "05807d",
+      R6: "0e00"
+    },
+    assets: []
+  }
+];
