+++ packages/core/src/builder/selector/strategies/accumulativeSelectionStrategy.spec.ts
@@ -1,18 +1,18 @@
 import { sumBy, utxoSum } from "@fleet-sdk/common";
-import { regularBoxesMock } from "_test-vectors";
+import { regularBoxes } from "_test-vectors";
 import { describe, expect, it } from "vitest";
 import { AccumulativeSelectionStrategy } from "./accumulativeSelectionStrategy";
 
 describe("Accumulative selection strategy", () => {
   it("Should return an empty array if empty target: { nanoErgs: 0 }", () => {
     const selector = new AccumulativeSelectionStrategy();
-    expect(selector.select(regularBoxesMock, { nanoErgs: 0n })).toEqual([]);
+    expect(selector.select(regularBoxes, { nanoErgs: 0n })).toEqual([]);
   });
 
   it("Should select inputs for nanoErgs only with target amount", () => {
     const selector = new AccumulativeSelectionStrategy();
     const target = { nanoErgs: 10000n };
-    const boxes = selector.select(regularBoxesMock, target);
+    const boxes = selector.select(regularBoxes, target);
 
     expect(boxes).toHaveLength(1);
     expect(sumBy(boxes, (x) => x.value)).toBeGreaterThanOrEqual(10000n);
@@ -20,15 +20,15 @@ describe("Accumulative selection strategy", () => {
 
   it("Should select all inputs with nanoErgs if no target amount is specified", () => {
     const selector = new AccumulativeSelectionStrategy();
-    const boxes = selector.select(regularBoxesMock, { nanoErgs: undefined });
+    const boxes = selector.select(regularBoxes, { nanoErgs: undefined });
 
-    expect(boxes).toHaveLength(regularBoxesMock.length);
+    expect(boxes).toHaveLength(regularBoxes.length);
   });
 
   it("Should select all inputs with a given token if no target amount is specified", () => {
     const selector = new AccumulativeSelectionStrategy();
 
-    const boxes = selector.select(regularBoxesMock, {
+    const boxes = selector.select(regularBoxes, {
       tokens: [{ tokenId: "007fd64d1ee54d78dd269c8930a38286caa28d3f29d27cadcb796418ab15c283" }]
     });
 
@@ -39,7 +39,7 @@ describe("Accumulative selection strategy", () => {
   it("Should select all inputs with a given token if no target amount is specified - multiple tokenIds", () => {
     const selector = new AccumulativeSelectionStrategy();
 
-    const boxes = selector.select(regularBoxesMock, {
+    const boxes = selector.select(regularBoxes, {
       tokens: [
         { tokenId: "007fd64d1ee54d78dd269c8930a38286caa28d3f29d27cadcb796418ab15c283" },
         { tokenId: "0cd8c9f416e5b1ca9f986a7f10a84191dfb85941619e49e53c0dc30ebf83324b" }
@@ -65,7 +65,7 @@ describe("Accumulative selection strategy", () => {
         }
       ]
     };
-    const inputs = selector.select(regularBoxesMock, target);
+    const inputs = selector.select(regularBoxes, target);
 
     expect(inputs).toHaveLength(1); // should try to reuse already selected inputs
     expect(sumBy(inputs, (x) => x.value)).toBeGreaterThanOrEqual(target.nanoErgs);
