+++ packages/core/src/builder/selector/strategies/cherryPickSelectionStrategy.spec.ts
@@ -1,18 +1,18 @@
 import { sumBy, utxoSum } from "@fleet-sdk/common";
-import { regularBoxesMock } from "_test-vectors";
+import { regularBoxes } from "_test-vectors";
 import { describe, expect, it } from "vitest";
 import { CherryPickSelectionStrategy } from "./cherryPickSelectionStrategy";
 
 describe("Cherry Pick selection strategy", () => {
   it("Should return an empty array if empty target: { nanoErgs: 0 }", () => {
     const selector = new CherryPickSelectionStrategy();
-    expect(selector.select(regularBoxesMock, { nanoErgs: 0n })).toEqual([]);
+    expect(selector.select(regularBoxes, { nanoErgs: 0n })).toEqual([]);
   });
 
   it("Should select only one input with no tokens in the input", () => {
     const selector = new CherryPickSelectionStrategy();
     const target = { nanoErgs: 10000n };
-    const inputs = selector.select(regularBoxesMock, target);
+    const inputs = selector.select(regularBoxes, target);
 
     expect(inputs).toHaveLength(1);
 
@@ -24,7 +24,7 @@ describe("Cherry Pick selection strategy", () => {
   it("Should select two inputs with no tokens in any of the selected inputs", () => {
     const selector = new CherryPickSelectionStrategy();
     const target = { nanoErgs: 67500000000n + 100000000n };
-    const inputs = selector.select(regularBoxesMock, target);
+    const inputs = selector.select(regularBoxes, target);
 
     expect(inputs).toHaveLength(2);
 
@@ -35,14 +35,14 @@ describe("Cherry Pick selection strategy", () => {
 
   it("Should select all inputs with nanoErgs if no target amount is specified", () => {
     const selector = new CherryPickSelectionStrategy();
-    const boxes = selector.select(regularBoxesMock, { nanoErgs: undefined });
+    const boxes = selector.select(regularBoxes, { nanoErgs: undefined });
 
-    expect(boxes).toHaveLength(regularBoxesMock.length);
+    expect(boxes).toHaveLength(regularBoxes.length);
   });
 
   it("Should select all inputs with a given token if no target amount is specified", () => {
     const selector = new CherryPickSelectionStrategy();
-    const boxes = selector.select(regularBoxesMock, {
+    const boxes = selector.select(regularBoxes, {
       tokens: [{ tokenId: "007fd64d1ee54d78dd269c8930a38286caa28d3f29d27cadcb796418ab15c283" }]
     });
 
@@ -64,7 +64,7 @@ describe("Cherry Pick selection strategy", () => {
       ]
     };
 
-    const inputs = selector.select(regularBoxesMock, target);
+    const inputs = selector.select(regularBoxes, target);
 
     expect(inputs).toHaveLength(2);
 
@@ -92,7 +92,7 @@ describe("Cherry Pick selection strategy", () => {
       ]
     };
 
-    const inputs = selector.select(regularBoxesMock, target);
+    const inputs = selector.select(regularBoxes, target);
 
     expect(inputs).toHaveLength(3);
