+++ packages/core/src/models/collections/inputsCollection.spec.ts
@@ -1,5 +1,5 @@
 import { first } from "@fleet-sdk/common";
-import { regularBoxesMock } from "_test-vectors";
+import { regularBoxes } from "_test-vectors";
 import { describe, expect, it } from "vitest";
 import { DuplicateInputError, NotFoundError } from "../../errors";
 import { InputsCollection } from "./inputsCollection";
@@ -11,13 +11,13 @@ describe("inputs collection", () => {
   });
 
   it("Should create a filled collection", () => {
-    const collection = new InputsCollection(regularBoxesMock);
-    expect(collection).toHaveLength(regularBoxesMock.length);
+    const collection = new InputsCollection(regularBoxes);
+    expect(collection).toHaveLength(regularBoxes.length);
   });
 
   it("Should add a single item", () => {
     const collection = new InputsCollection();
-    const box = first(regularBoxesMock);
+    const box = first(regularBoxes);
     collection.add(box);
 
     expect(collection).toHaveLength(1);
@@ -26,10 +26,10 @@ describe("inputs collection", () => {
 
   it("Should append items", () => {
     const collection = new InputsCollection();
-    collection.add(regularBoxesMock);
+    collection.add(regularBoxes);
 
-    expect(collection).toHaveLength(regularBoxesMock.length);
-    expect(collection.toArray()).toEqual(regularBoxesMock);
+    expect(collection).toHaveLength(regularBoxes.length);
+    expect(collection.toArray()).toEqual(regularBoxes);
   });
 
   it("Should add a multiple items and map properly", () => {
@@ -37,7 +37,7 @@ describe("inputs collection", () => {
 
     // include boxes with amounts as string
     collection.add(
-      regularBoxesMock.map((box) => {
+      regularBoxes.map((box) => {
         return {
           ...box,
           value: box.value.toString(),
@@ -48,38 +48,38 @@ describe("inputs collection", () => {
       })
     );
 
-    expect(collection).toHaveLength(regularBoxesMock.length);
-    expect(collection.toArray()).toEqual(regularBoxesMock);
+    expect(collection).toHaveLength(regularBoxes.length);
+    expect(collection.toArray()).toEqual(regularBoxes);
   });
 
   it("Should throw if box is already included", () => {
-    const collection = new InputsCollection(regularBoxesMock);
+    const collection = new InputsCollection(regularBoxes);
     expect(() => {
-      collection.add(first(regularBoxesMock));
+      collection.add(first(regularBoxes));
     }).toThrow(DuplicateInputError);
   });
 
   it("Should remove by boxId", () => {
-    const collection = new InputsCollection(regularBoxesMock);
-    const boxId = first(regularBoxesMock).boxId;
+    const collection = new InputsCollection(regularBoxes);
+    const boxId = first(regularBoxes).boxId;
     collection.remove(boxId);
 
-    expect(collection).toHaveLength(regularBoxesMock.length - 1);
+    expect(collection).toHaveLength(regularBoxes.length - 1);
     expect(collection.toArray().find((x) => x.boxId === boxId)).toBeFalsy();
   });
 
   it("Should remove by index", () => {
-    const collection = new InputsCollection(regularBoxesMock);
-    const boxId = first(regularBoxesMock).boxId;
+    const collection = new InputsCollection(regularBoxes);
+    const boxId = first(regularBoxes).boxId;
     collection.remove(0);
 
-    expect(collection).toHaveLength(regularBoxesMock.length - 1);
+    expect(collection).toHaveLength(regularBoxes.length - 1);
     expect(collection.toArray().find((x) => x.boxId === boxId)).toBeFalsy();
   });
 
   it("Should throw if not found", () => {
-    const collection = new InputsCollection(first(regularBoxesMock));
-    const boxId = regularBoxesMock[1].boxId;
+    const collection = new InputsCollection(first(regularBoxes));
+    const boxId = regularBoxes[1].boxId;
 
     expect(() => {
       collection.remove(boxId);
@@ -87,14 +87,14 @@ describe("inputs collection", () => {
   });
 
   it("Should throw if not found", () => {
-    const collection = new InputsCollection(regularBoxesMock);
+    const collection = new InputsCollection(regularBoxes);
 
     expect(() => {
       collection.remove(-1);
     }).toThrow(RangeError);
 
     expect(() => {
-      collection.remove(regularBoxesMock.length);
+      collection.remove(regularBoxes.length);
     }).toThrow(RangeError);
 
     expect(() => {
