+++ packages/serializer/src/_test-vectors/constantVectors.ts
