+++ plugins/babel-fees/src/utils.spec.ts
@@ -1,8 +1,7 @@
 import { first } from "@fleet-sdk/common";
+import { babelBoxesMock, regularBoxes } from "_test-vectors";
 import { describe, expect, it } from "vitest";
-import { babelBoxesMock } from "./tests/babelBoxesMock";
-import { regularBoxesMock } from "./tests/regularBoxesMock";
-import { babelContractTestVectors } from "./tests/testVectors";
+import { babelContractTestVectors } from "./_test-vectors/babelContractsVectors";
 import {
   buildBabelContract,
   extractTokenIdFromBabelContract,
@@ -17,7 +16,7 @@ describe("getTokenPrice()", () => {
   });
 
   it("Should fail with an invalid Babel Box", () => {
-    for (const regularBox of regularBoxesMock) {
+    for (const regularBox of regularBoxes) {
       expect(() => {
         getTokenPrice(regularBox);
       }).toThrow();
